requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
solana>=0.34.2
solders>=0.21.0,<0.28.0
base58>=2.1.1
cryptography>=41.0.7
websockets>=12.0
aiohttp>=3.9.1
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
websocket-client>=1.6.4
pydantic>=2.5.3
construct>=2.10.70
python-telegram-bot>=21.7
httpx>=0.27.0,<0.28.0

